# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Antojados API
  version: 1.0.0
  description: API REST de AntojadosMX - Auth, Feed, Places, Posts, Social, Biz, Rewards, Media, Sync, Analytics, Equipo, GT Panel.
servers:
  - url: https://api.antojadosmx.mx
    description: Produccion (Cloudflare HTTPS)
  - url: http://185.187.235.253:8010
    description: Contabo (IP directa - fallback)
...
paths:
...
  /api/v1/antojados/posts/{post_id}:
    get:
      tags: [Posts]
      summary: Obtener post
      parameters: [{in: path, name: post_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
    delete:
      tags: [Posts]
      summary: Eliminar post
      parameters: [{in: path, name: post_id, required: true, schema: {type: string}}]
      responses:
        "204": { description: No Content }
  /api/v1/antojados/posts/{post_id}/like:
    post:
      tags: [Social]
      summary: Dar like a post
      parameters: [{in: path, name: post_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
    delete:
      tags: [Social]
      summary: Quitar like de post
      parameters: [{in: path, name: post_id, required: true, schema: {type: string}}]
      responses:
        "204": { description: No Content }
  /api/v1/antojados/posts/{post_id}/comments:
    post:
      tags: [Social]
      summary: Comentar post
      parameters: [{in: path, name: post_id, required: true, schema: {type: string}}]
      responses:
        "201": { description: Created }
    get:
      tags: [Social]
      summary: Comentarios de post
      parameters: [{in: path, name: post_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/posts/{post_id}/ratings:
    post:
      tags: [Social]
      summary: Calificar post
      parameters: [{in: path, name: post_id, required: true, schema: {type: string}}]
      responses:
        "201": { description: Created }
...
  /api/v1/antojados/social/following/{user_id}:
    get:
      tags: [Social]
      summary: Usuarios que sigue
      parameters: [{in: path, name: user_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/social/followers/{user_id}:
    get:
      tags: [Social]
      summary: Seguidores del usuario
      parameters: [{in: path, name: user_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/social/saves/{user_id}:
    get:
      tags: [Social]
      summary: Guardados del usuario
      parameters: [{in: path, name: user_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/social/saves/{user_id}/feed:
    get:
      tags: [Social]
      summary: Feed de guardados del usuario
      parameters: [{in: path, name: user_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/rewards/campaigns/{campaign_id}:
    get:
      tags: [Rewards]
      summary: Detalle de campana
      parameters: [{in: path, name: campaign_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/rewards/redemptions/{user_id}:
    get:
      tags: [Rewards]
      summary: Historial de canjes del usuario
      parameters: [{in: path, name: user_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...